+++ src/runtime/python_runtime/CMakeLists.txt
@@ -13,13 +13,9 @@ set(CUR_TARGET_NAME ${CUR_SUPERIOR_NAMESPACE_UNDERLINE}_${CUR_DIR})
 set(CUR_TARGET_ALIAS_NAME ${CUR_SUPERIOR_NAMESPACE}::${CUR_DIR})
 
 # Set file collection
-file(GLOB_RECURSE src CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.cc)
+file(GLOB_RECURSE src CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/python_runtime_main.cc)
 file(GLOB_RECURSE py_files CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.py)
-file(GLOB_RECURSE ros2_files CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/ros2_type_support_utils.cc)
-
-if(NOT AIMRT_BUILD_WITH_ROS2)
-  list(REMOVE_ITEM src ${ros2_files})
-endif()
+file(GLOB_RECURSE ros2_files CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/ros2_type_support_utils.cc ${CMAKE_CURRENT_SOURCE_DIR}/python_runtime_ros2.cc)
 
 # Add target
 pybind11_add_module(${CUR_TARGET_NAME} SHARED)
@@ -33,13 +29,21 @@ target_link_libraries(
   ${CUR_TARGET_NAME}
   PRIVATE $<LINK_LIBRARY:WHOLE_ARCHIVE,aimrt::runtime::core>)
 
-# Link ros2 interface if building with ros2
-if(AIMRT_BUILD_WITH_ROS2)
-  target_link_libraries(${CUR_TARGET_NAME} PRIVATE aimrt::interface::aimrt_module_ros2_interface)
-  target_compile_definitions(${CUR_TARGET_NAME} PRIVATE AIMRT_BUILD_WITH_ROS2)
-endif()
-
 # Set misc of target
 set_target_properties(${CUR_TARGET_NAME} PROPERTIES OUTPUT_NAME "aimrt_python_runtime")
 
 set_property(TARGET ${CUR_TARGET_NAME} PROPERTY PY_FILES ${py_files})
+
+if(AIMRT_BUILD_WITH_ROS2)
+  pybind11_add_module(${CUR_TARGET_NAME}_ros2 SHARED)
+  add_library(${CUR_TARGET_ALIAS_NAME}_ros2 ALIAS ${CUR_TARGET_NAME}_ros2)
+
+  target_sources(${CUR_TARGET_NAME}_ros2 PRIVATE ${ros2_files})
+
+  target_link_libraries(
+    ${CUR_TARGET_NAME}_ros2
+    PRIVATE $<LINK_LIBRARY:WHOLE_ARCHIVE,aimrt::runtime::core>
+    aimrt::interface::aimrt_module_ros2_interface)
+
+  set_target_properties(${CUR_TARGET_NAME}_ros2 PROPERTIES OUTPUT_NAME "aimrt_python_runtime_ros2")
+endif()
