+++ src/runtime/python_runtime/python_runtime_main.cc
@@ -14,10 +14,6 @@
 #include "python_runtime/export_pb_type_support.h"
 #include "python_runtime/export_rpc.h"
 
-#ifdef AIMRT_BUILD_WITH_ROS2
-  #include "python_runtime/export_ros2_type_support.h"
-#endif
-
 using namespace aimrt::runtime::python_runtime;
 
 PYBIND11_MODULE(aimrt_python_runtime, m) {
@@ -29,9 +25,6 @@ PYBIND11_MODULE(aimrt_python_runtime, m) {
 
   // type support
   ExportPbTypeSupport(m);
-#ifdef AIMRT_BUILD_WITH_ROS2
-  ExportRos2TypeSupport(m);
-#endif
 
   // core handle
   ExportModuleInfo(m);
