+++ src/tools/package_aimrt_py/CMakeLists.txt
@@ -16,6 +16,15 @@ add_custom_target(
   COMMENT "Copying python runtime files"
   DEPENDS aimrt::runtime::python_runtime)
 
+if(AIMRT_BUILD_WITH_ROS2)
+  add_custom_target(
+    copy_ros2_python_runtime
+    COMMAND ${CMAKE_COMMAND} -E make_directory ${AIMRT_PY_PKG_DIR}/aimrt_py/
+    COMMAND ${CMAKE_COMMAND} -E copy_if_different $<TARGET_FILE:aimrt::runtime::python_runtime_ros2> ${AIMRT_PY_PKG_DIR}/aimrt_py/
+    COMMENT "Copying python runtime ros2 files"
+    DEPENDS aimrt::runtime::python_runtime_ros2)
+endif()
+
 add_custom_target(
   copy_ros2_plugin_proto_files
   COMMAND ${CMAKE_COMMAND} -E make_directory ${AIMRT_PY_PKG_DIR}/aimrt_py/
