+++ src/tools/protoc_plugin_py_gen_aimrt_py_rpc/protoc_plugin_py_gen_aimrt_py_rpc.py
@@ -22,7 +22,7 @@ from typing import overload
 
 import aimrt_py
 import google.protobuf
-from . import {{py_package_name}}
+import {{py_package_name}}
 {{pyfile_import_dependency_py_package}}
 
 {{for service begin}}
